Add files via upload
Here in this, we are going to reverse the array and updating the original array. To do this, we can  use two approaches. First, we can create another array of the same size and then we are going to transfer elements from the original array to the created array in reverse order and then we are going to reassign the same values back in the original array.
Second approach is a two pointer approach in which we just reversed the original array without use of extra space for array.